Create Grade Book
Create a Grade Book that uses a two-dimensional array to store the grades of a number of students on
multiple exams. Each row of the array represents a single student’s grades for the entire course, and
each column represents the grades of all the students who took a particular exam.
Required Output:
Welcome to the grade book for Software Construction &amp; Development!
The grades are:
Name            Exam-I   Exam-II   Exam-III  Average
Student 1
Student 2
Student 3
Student 4
Student 5
Student 6
Student 7
Student 8
Student 9
Student 10
The lowest grade in the grade book is =?
The highest grade in the grade book is =?
Grade Distribution based on Average:
Grades # of Students
Below 50 2
51 – 60
61 – 70 3
71 – 80 1
81 - 90 3
91 - 100 
Note There Are 10 Students and 3 Exam.